Regexp improvements, also added .gitignore
Changed regexps in fic-search-re to not mandate that the beginning and
end of "TODO", "FIXME", etc. be on symbol boundaries, which was preventing
those from getting highlighted if followed by a symbol other than whitespace.
A better way to fix this would be to have a new variable for the
character class allowed only as the character immediately following the
symbol.